Struct iowrap::Ignore

pub struct Ignore {}

Ignore all IO requests made on this object.

For Write, that means reporting success always, but not doing anything.

For Read and Seek, that means acting like an immutable empty file.

This may confuse things which use Write and Seek together, as writing to an Ignore does not advance its seek position, which is not what would happen with a File or Cursor or similar. The Seek implementation ignores out of bound seeks, negative seeks, etc., in the hope of maybe mitigating this a bit.

Example

use iowrap::Ignore;
assert_eq!(0, io::copy(&mut Ignore::new(), &mut Ignore::new()).unwrap());

Implementations

impl Ignore

pub fn new() -> Self